refactor(app): add LibroAppView
