+++ packages/libro-app/src/index.less
@@ -133,3 +133,7 @@
     }
   }
 }
+
+.libro-app-viewer-container {
+  height: 100%;
+}
